04-weather-agent/pyproject.toml
[project]
name = "weather-agent"
version = "0.1.0"
description = "AI Weather Agent with manual tools protocol implementation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "openai>=1.60.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
